moved Report components to a folder
