+++ oya-web/src/report/PieReportTotal.tsx
@@ -3,7 +3,7 @@ import {
   Tooltip, ResponsiveContainer, PieChart, Pie, PieLabelRenderProps,
 } from 'recharts';
 import { ActivityTotalReport } from './reportUtils';
-import { Activity } from './apiService';
+import { Activity } from '../apiService';
 import RechartsTooltip from './RechartsTooltip';
 import { useReportContext } from './ReportProvider';
