+++ oya-web/src/report/TreemapReportRechart.tsx
@@ -1,7 +1,7 @@
 import { Box } from '@mui/material';
 import { Treemap, Tooltip, ResponsiveContainer } from 'recharts';
 import { ActivityTotalReport } from './reportUtils';
-import { Activity } from './apiService';
+import { Activity } from '../apiService';
 import RechartsTooltip from './RechartsTooltip';
 import { useReportContext } from './ReportProvider';
